stream_base: move Writev/WriteString
