+++ src/stream_base.h
@@ -87,16 +87,11 @@ class StreamBase : public StreamResource {
   static void AddMethods(Environment* env,
                          v8::Handle<v8::FunctionTemplate> target);
 
-  virtual bool IsAlive() = 0;
-  virtual int GetFD() = 0;
+  virtual bool IsAlive() const = 0;
+  virtual int GetFD() const = 0;
 
   virtual int ReadStart() = 0;
   virtual int ReadStop() = 0;
-  virtual int Writev(v8::Local<v8::Object> req, v8::Local<v8::Array> bufs) = 0;
-  virtual int WriteString(v8::Local<v8::Object> req,
-                          v8::Local<v8::String> str,
-                          enum encoding enc,
-                          v8::Local<v8::Object> handle) = 0;
   virtual int SetBlocking(bool enable) = 0;
 
   inline bool IsConsumed() const { return consumed_; }
@@ -109,6 +104,7 @@ class StreamBase : public StreamResource {
   virtual ~StreamBase() = default;
 
   virtual v8::Local<v8::Object> GetObject() = 0;
+  virtual bool IsIPCPipe() const = 0;
 
   // Libuv callbacks
   static void AfterShutdown(uv_shutdown_t* req, int status);
@@ -120,7 +116,7 @@ class StreamBase : public StreamResource {
   int Shutdown(const v8::FunctionCallbackInfo<v8::Value>& args);
   int Writev(const v8::FunctionCallbackInfo<v8::Value>& args);
   int WriteBuffer(const v8::FunctionCallbackInfo<v8::Value>& args);
-  template <enum encoding Encoding>
+  template <enum encoding enc>
   int WriteString(const v8::FunctionCallbackInfo<v8::Value>& args);
   int SetBlocking(const v8::FunctionCallbackInfo<v8::Value>& args);
